last day

+++ src/main/java/com/ra/hotel_booking/model/entity/Customer.java
@@ -16,7 +16,7 @@ import java.util.Set;
 public class Customer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long customer_id;
+    private int customer_id;
     @Column(name = "first_name", length = 50)
     private String first_name;
     @Column(name = "last_name", length = 50)
